move env stuff to env.rs
